Skip docs sync PR with no changes.

+++ .expeditor/generate-automate-api-docs.sh
@@ -12,12 +12,16 @@ pushd /go/src/github.com/chef/automate/components/automate-chef-io
   make sync_swagger_files
 popd
 
-git add components/automate-chef-io/data/docs/api_chef_automate
+if [[ `git status --porcelain` ]]; then
+  git add components/automate-chef-io/data/docs/api_chef_automate
 
-git commit --message "Sync swagger files for Automate docs." --message "This pull request was triggered automatically via Expeditor." --message "This change falls under the obvious fix policy so no Developer Certificate of Origin (DCO) sign-off is required."
+  git commit --message "Sync swagger files for Automate docs." --message "This pull request was triggered automatically via Expeditor." --message "This change falls under the obvious fix policy so no Developer Certificate of Origin (DCO) sign-off is required."
 
-open_pull_request
+  open_pull_request
 
-git checkout -
-git clean -fxd
-git branch -D "$branch"
+  git checkout -
+  git clean -fxd
+  git branch -D "$branch"
+else
+  echo "No changed files detected, skipping sync PR."
+fi
